moved `TooltipVisibility` to new file
